+++ src/lib/connection/XMPPServiceHandler/TasksEventHandler.ts
@@ -217,6 +217,9 @@ class TasksEventHandler extends GenericHandler {
                     case "userpassword":
                         // treated in conversationEventHandler
                         break;
+                    case "customStatus":
+                        // treated in presentationEventHandler
+                        break;
                     default:
                         that._logger.log(that.ERROR, LOG_ID + "(onManagementMessageReceived) unmanaged management message node " + node.getName());
                         break;