+++ src/lib/connection/XMPPServiceHandler/alertEventHandler.ts
@@ -410,6 +410,14 @@ class AlertEventHandler extends GenericHandler {
                     case "todo":
                         // treated in tasksEventHandler
                         break;
+                    case "no-store":
+                        // treated in conversationEventHandler
+                        break;
+                    case "userpassword":
+                        // treated in conversationEventHandler
+                        break;
+                    case "customStatus":
+                    // treated in presentationEventHandler
                     default:
                         that._logger.log(that.ERROR, LOG_ID + "(onManagementMessageReceived) unmanaged management message node " + node.getName());
                         break;