+++ src/lib/connection/XMPPServiceHandler/invitationEventHandler.ts
@@ -146,6 +146,8 @@ class InvitationEventHandler extends GenericHandler {
                     case "userpassword":
                         // treated in conversationEventHandler
                         break;
+                    case "customStatus":
+                    // treated in presentationEventHandler
                     default:
                         that._logger.log(that.ERROR, LOG_ID + "(onManagementMessageReceived) unmanaged management message node " + node.getName());
                         break;