+++ src/lib/connection/XMPPServiceHandler/webinarEventHandler.ts
@@ -188,6 +188,8 @@ class WebinarEventHandler extends GenericHandler {
                     case "userpassword":
                         // treated in conversationEventHandler
                         break;
+                    case "customStatus":
+                    // treated in presentationEventHandler
                     default:
                         that._logger.log(that.ERROR, LOG_ID + "(onManagementMessageReceived) unmanaged management message node " + node.getName());
                         break;